Sanctum impl for authentication and Gate && policies implementationsfor authorization

+++ laravel11/event-management/app/Http/Controllers/Api/AttendeeController.php
@@ -7,13 +7,21 @@ use App\Http\Resources\AttendeeResource;
 use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
-    use CanLoadRelationships;
+    use CanLoadRelationships, AuthorizesRequests;
 
     private array $relations = ['user'];
+
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum')->except(['index', 'show']);
+        $this->authorizeResource(Attendee::class, 'attendee');
+    }
     /**
      * Display a listing of the resource.
      */
@@ -57,8 +65,12 @@ class AttendeeController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $event, Attendee $attendee)
+    public function destroy(Event $event, Attendee $attendee)
     {
+        // Gate::authorize('delete-attendee', $attendee);
+        // OR
+        Gate::authorize('delete-attendee', [$event, $attendee]);
+
         $attendee->delete();
 
         return response(status: 204);
